Moved useEffect in telemetry table to second screen so reddit feed can be updated each lap
